+++ NEWS.md
@@ -2,11 +2,9 @@
 
 * Removed tcltk, rlist dependencies (holdovers from older code)
 
-* Updated some of the innards to make post-processing more efficient
+* Streamlined some things under the hood to make post-processing more efficient in eloratingopt & eloratingfixed
 
 * EloOrdinal column in output now stored as an integer object for small space saving.
-
-* Still need to update the join; update post-processing in eloratingfixed
 
 # EloOptimized 0.3.0